+++ include/core/o3_cScrBuf.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2010 Javeline BV
+ * Copyright (C) 2010 Ajax.org BV
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -81,7 +81,7 @@ struct cScrBuf : cScr, iBuf {
         o3_trace3 trace;
 
         if ((size_t) index < m_buf.size())
-            return ((uint8_t*) m_buf.ptr())[index] = b;
+            return (int) (((uint8_t*) m_buf.ptr())[index] = (uint8_t)b);
         return 0; 
     }
